requirements.txt
pandas==2.2.2
Pillow==10.3.0
streamlit==1.34.0
ultralytics==8.2.8
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
requests==2.31.0
